+++ modules/Content/views/singleton/item.php
@@ -65,6 +65,7 @@
                         </kiss-card>
                     </div>
 
+                    <?php if ($this->helper('acl')->isAllowed("content/{$model['name']}/publish")): ?>
                     <div class="kiss-margin">
 
                         <div class="kiss-text-caption kiss-size-xsmall kiss-text-bold">{{ t('State') }}</div>
@@ -78,6 +79,7 @@
                         </div>
 
                     </div>
+                    <?php endif ?>
 
                     <div class="kiss-margin" v-if="hasLocales">
 
@@ -159,6 +161,7 @@
                 </kiss-content>
             </kiss-popoutmenu>
 
+            <?php if ($this->helper('acl')->isAllowed("content/{$model['name']}/publish")): ?>
             <kiss-popoutmenu id="model-item-menu-state">
                 <kiss-content>
                     <kiss-navlist class="kiss-margin-small">
@@ -182,6 +185,7 @@
                     </kiss-navlist>
                 </kiss-content>
             </kiss-popoutmenu>
+            <?php endif ?>
 
             <kiss-popoutmenu id="model-item-preview-links" v-if="model.preview && model.preview.length">
                 <kiss-content>