+++ modules/App/assets/dialog/app-search.js
@@ -114,9 +114,12 @@ export default {
                 <div class="kiss-color-muted kiss-margin-small-right">
                     <kiss-svg :src="$base('system:assets/icons/search.svg')" width="30"><canvas width="30" height="30"></canvas></kiss-svg>
                 </div>
-                <div class="kiss-flex-1">
+                <div class="kiss-flex-1 kiss-margin-small-right">
                     <input class="kiss-input kiss-input-blank kiss-form-large kiss-width-1-1" :class="{'kiss-disabled': loading}" type="text" v-model="search" :placeholder="t('Search...')" :disabled="loading" @keydown="keydown" @input="selected=null" ref="searchInput" style="font-size:30px;padding:0;">
                 </div>
+                <a class="kiss-color-muted" kiss-dialog-close>
+                    <kiss-svg :src="$base('system:assets/icons/close.svg')" width="30"><canvas width="30" height="30"></canvas></kiss-svg>
+                </a>
             </div>
 
             <kiss-card class="kiss-color-muted kiss-size-large kiss-align-center kiss-padding" v-if="loading">