+++ modules/App/layouts/app.php
@@ -38,7 +38,7 @@
                         <ul>
                             <li class="<?=($this->request->route == '/') ? 'active':''?>">
                                 <a href="<?=$this->route('/')?>" aria-label="<?=t('Dashboard')?>" kiss-tooltip="right">
-                                    <kiss-svg src="<?=$this->base('app:icon.svg')?>" width="25" height="25"></kiss-svg>
+                                    <kiss-svg src="<?=$this->base('app:icon.svg')?>" width="25" height="25"><canvas width="25" height="25"></canvas></kiss-svg>
                                 </a>
                             </li>
                             <?php foreach ($this->helper('menus')->menu('modules', true) as $group => $links): ?>
@@ -48,12 +48,18 @@
                                 <?php foreach ($links as $link): ?>
                                     <li class="<?=(strpos($this->request->route, $link['route']) === 0) ? 'active':''?>">
                                         <a href="<?=$this->route($link['route'])?>" aria-label="<?=t($link['label'])?>" kiss-tooltip="right">
-                                            <kiss-svg src="<?=$this->base($link['icon'])?>" width="25" height="25"></kiss-svg>
+                                            <kiss-svg src="<?=$this->base($link['icon'])?>" width="25" height="25"><canvas width="25" height="25"></canvas></kiss-svg>
                                         </a>
                                     </li>
                                 <?php endforeach ?>
 
                             <?php endforeach ?>
+                            <li class="kiss-nav-divider"></li>
+                            <li>
+                                <a class="kiss-flex kiss-flex-center" aria-label="<?=t('Search')?>" kiss-tooltip="right" app-search>
+                                    <kiss-svg src="<?=$this->base('system:assets/icons/search.svg')?>" width="25"><canvas width="25" height="25"></canvas></kiss-svg>
+                                </a>
+                            </li>
                         </ul>
                     </kiss-navlist>
                 </div>