+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 
 WIP
 
-- General cleanup + fiy typos
-- Fix typo field-contentItemtLink -> !!! Need to re-assign field type contentItemLink !!!
+- 🚨 Fix typo field-contentItemtLink -> Need to re-assign field type contentItemLink 🚨
+- General cleanup + fix typos
+- Add content tree items reorder permission
 
 ## 2.1.1 (2022-08-02)