+++ CHANGELOG.md
@@ -10,6 +10,7 @@ WIP
 - Show more detailed error message if saving user settings fail
 - Hide publish state configuration ui if missing permission to mange publish state
 - Fix selected item link of tree model
+- Fix missing assets folders permission checks
 
 ## 2.1.1 (2022-08-02)